Simplify DoIt's build script.
Buildscr is still under the impression that we're building a generic
'listener' module into a set of related programs, and I accepted some
years ago that this never actually happened and DoIt is the only
program that this framework will ever be used for.

Accordingly, chop out all the generic stuff in the build script, which
will make it less painful to start modifying in other ways in a
moment.

[originally from svn r10245]