Allow DoIt-launched programs to open in the foreground.
The idea of this comes from a patch by Russell Davis in a fork of
DoIt, who in turn cites a Google Groups thread discussing why it
works. However, I've replaced the actual patch completely (rewritten
the comment in my own words, and rewritten the code to use a
completely different API call that isn't deprecated).

The original patch is not based on DoIt's own git history at all, but
on an independent import into git of a DoIt distribution archive, so I
can't cite the original commit as a parent. At the time of writing
this, it appears at https://github.com/russelldavis/DoIt as commit
d00073cad03542ae1ecbcede49056ebff7b637b0.

Thanks also to Chris Nappi for bringing both this issue and Russell
Davis's work to my attention.