Preliminary check of the SetDirectory argument.
Another very common cause of incomprehensible DoIt server errors is
that a CreateProcess call fails because the pathname given as its cwd
is invalid (again, often due to non-translation on the client side).
Now when we receive a SetDirectory prefix call, we check whether its
argument _is_ a directory, and if not, we construct a sensible error
message that will actually say that instead of just reporting what
went wrong in the subsequent CreateProcess.

Because some of the process-creation commands expect the command's
output to be sent before the error message, we can't just return that
error message as soon as the bogus SetDirectory is received; instead,
we stash it as a 'setup error', and return that in place of the
follow-up error _after_ whatever else the protocol requires us to
send.