Revamp reporting of network protocol errors.
doitlib.c's main handler function for incoming data now returns a
textual error message in place of a simple -1, so that the client can
actually print some useful indication of _what_ went wrong in the
protocol. The other job of that return value - giving the current
amount of buffered plaintext - now belongs to a separate function
doit_buffered(), which is more sensible anyway because now it's
impossible to get the two kinds of return value confused (which I
think happened in at least one place in the old code).