Detect and diagnose shared-secret mismatch at connection start.
This is a protocol change, the first in a very long time (so users
will need to upgrade both client and server in sync, per the warning
I've had on the web page all along). I've started by introducing a
protocol version field at the very start of the connection, so that
future version mismatches can be reliably diagnosed. Then each side
sends the same nonce as before, but accompanies it with a (two-layer)
hash of the nonce and the shared secret, so that if the shared secrets
simply aren't the same, we can diagnose that right at the start of the
connection and give an error message that provides some kind of hint
about what might have gone wrong.