Fix stray 1997 in About box copyright notice.
Every _other_ copy of the copyright notice in this source base, along
with the git history, says I actually started this project in 2000,
not 1997. I think the 1997 must have been an accidental copy-paste
from a previous project that used a similar piece of Windows .rc file
for its About box, and I pasted in the revised actual licence but
forgot to adjust the extra copy of the copyright date.