Usefully diagnose absence of a shared secret.
If the DoIt server can't load its secret file, it will now put up a
message box, and then try again on the next connection attempt (even
if not in -r mode).

Also, I've added a protocol feature to report a 'service unavailable'
message to the client, so that there's at least some user-friendliness
if you find your server is in this state.