Improve uniqueness of protocol nonces.
I was a bit unhappy about time(NULL) being the form of timestamp used
for uniqueness. Now we use a high-resolution timer on both Windows and
Linux, and also bake in a few other local things like process ids at
both ends rather than just on Windows.

To make this easier, I revamped the API for retrieving nonce preimage
data so that it's now called from doitlib.c. An effect of this is that
it would be more hassle to make the nonce preimage depend on the
client's network connection source address, which was a thing I _was_
previously doing on the server side. But I don't think that's
particularly important, and I've replaced it with an incrementing
counter which should be enough all by itself to ensure that two nearby
connections made to the same server process won't manage to give the
same nonce.