Crypto: switch the protocol to encrypt-then-MAC.
This follows what now seems to be standard best practice: when
ciphertext is received, the very first thing that happens is that its
MAC gets verified, and it is not entrusted to any other part of the
code base until that's done. That automatically prevents a variety of
attacks; in particular, decryption oracle attacks involving modifying
the ciphertext in transit and observing the resulting behaviour of the
receiving implementation all now stop working because the behaviour in
question is invariably 'complain about MAC failure'.

Encrypt-then-MAC doesn't work completely trivially with the SSH-style
packet format I was using: if the packet length is inside the first
cipher block, then you can't find it out until after you've decrypted
that block, but if you can't decrypt anything until you've MACed the
whole packet for which you need to know its length, then there's an
obvious dependency cycle. To fix that I've broken up any packet longer
than one cipher block into an 'initial packet' one block in length and
a 'followup packet' containing everything else, and they're MACed
separately, so you can verify the MAC on the first cipher block, then
decrypt it to find the packet length, _then_ read, MAC and decrypt the
followup packet if there is one.