Crypto: switch to using AES in counter mode, not CBC.
This is another thing that people seem to have settled on as good
practice these days. It effectively turns your block cipher into a
stream cipher, with a keystream derived from encrypting an
incrementing sequence of 128-bit integers which is then XORed into the
plaintext. So encryption and decryption are the same operations.

The risks of a stream cipher are that an active attacker can make
extremely controlled changes to the plaintext by modifying the
ciphertext (so I couldn't switch to this mode until the MAC was being
checked on all ciphertext before depending on it in _any_ other way,
including decoding the length field out of the packet header), and
that if you reuse the same piece of keystream then a listener can
trivially derive the XOR of two pieces of plaintext (but we never do,
because we only ever regenerate the counter / IV when we also generate
a fresh AES key).

On the plus side, this removes the possibility of a repeated cipher
block. In CBC mode an n-bit block cipher would expect a repeat after
only about 2^{n/2} blocks, due to birthday-paradox considerations; in
counter mode, a repeated cipher block absolutely _cannot_ occur until
the counter has cycled round all 2^n possible block values.